style: User 관련 컴포넌트 분해 및 파일 이동 처리
