+++ src/components/Manager/User/UserDetailModal.tsx
@@ -1,0 +1,197 @@
+import { Button } from "@mui/material";
+import { DataGrid } from "@mui/x-data-grid";
+import { useRef } from "react";
+import { styled } from "styled-components";
+
+import { StyledModal } from "@/components/Manager/DashboardItems";
+import {
+  companyAddressColumns,
+  personalInfoColumns1,
+  personalInfoColumns2,
+  productListColumns,
+  shippingAddressColumns,
+} from "@/components/Manager/User/const";
+import { handleUserInfoForOrder } from "@/components/Manager/User/util";
+import { OrdererInfo } from "@/const";
+
+const StyledModalContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  width: 740px;
+  height: 650px;
+  background-color: var(--color-white);
+  overflow: scroll;
+  padding: 30px;
+
+  > div {
+    width: 100%;
+  }
+
+  h4 {
+    margin: 10px 0;
+    color: var(--color-h4-secondary);
+  }
+`;
+
+const FlexDiv = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+`;
+
+const StyledCopyButton = styled(Button)`
+  height: 30px;
+`;
+
+const UserDetailModal = ({
+  isModalOpen,
+  onModalClose,
+  modalUserInfoData,
+}: {
+  isModalOpen: boolean;
+  onModalClose: () => void;
+  modalUserInfoData: OrdererInfo;
+}) => {
+  const {
+    userId,
+    submissionTime,
+    productLengthUnit,
+    hopeProducts,
+    personalInfo: {
+      name,
+      companyName,
+      businessType,
+      contactInfo: { email, phoneNumber },
+      companyAddress,
+      shippingAddress,
+    },
+  } = modalUserInfoData;
+
+  const orderRows = handleUserInfoForOrder(hopeProducts);
+  const copyButtonRef = useRef<HTMLButtonElement>(null);
+
+  const handleCopyClipBoard = async () => {
+    const text = hopeProducts.map((product) => product.productId).join("\n");
+    try {
+      await navigator.clipboard.writeText(text);
+      if (copyButtonRef.current) {
+        copyButtonRef.current.textContent = "복사됨";
+      }
+    } catch (e) {
+      alert("복사에 실패하였습니다");
+    }
+  };
+
+  return (
+    <StyledModal open={isModalOpen} onClose={onModalClose}>
+      <StyledModalContainer>
+        <h2>User Info - {submissionTime ?? ""}</h2>
+        <h4>Personal Info</h4>
+        <div>
+          <DataGrid
+            rows={[
+              {
+                id: userId,
+                name,
+                company: companyName,
+              },
+            ]}
+            columns={personalInfoColumns1}
+            autoHeight
+            hideFooter
+            density="compact"
+            sx={{
+              borderLeft: "none",
+              borderRight: "none",
+            }}
+          />
+        </div>
+        <div>
+          <DataGrid
+            getRowId={() => userId}
+            rows={[
+              {
+                email,
+                contactNumber: `${phoneNumber.countryCode}${phoneNumber.number}`,
+                businessType,
+              },
+            ]}
+            columns={personalInfoColumns2}
+            autoHeight
+            hideFooter
+            density="compact"
+            sx={{
+              borderLeft: "none",
+              borderRight: "none",
+            }}
+          />
+        </div>
+        <h4>Company Address</h4>
+        <div>
+          <DataGrid
+            getRowId={() => userId}
+            rows={[
+              {
+                postalCode: companyAddress.postalCode,
+                address: companyAddress.address,
+                detailAddress: companyAddress.detailAddress,
+              },
+            ]}
+            columns={companyAddressColumns}
+            autoHeight
+            hideFooter
+            density="compact"
+            sx={{
+              borderLeft: "none",
+              borderRight: "none",
+            }}
+          />
+        </div>
+        <h4>Shipping Address</h4>
+        <div>
+          <DataGrid
+            getRowId={() => userId}
+            rows={[
+              {
+                postalCode: shippingAddress.postalCode,
+                address: shippingAddress.address,
+                detailAddress: shippingAddress.detailAddress,
+              },
+            ]}
+            columns={shippingAddressColumns}
+            autoHeight
+            hideFooter
+            density="compact"
+            sx={{
+              borderLeft: "none",
+              borderRight: "none",
+            }}
+          />
+        </div>
+
+        <FlexDiv>
+          <h2>Order - {productLengthUnit.toLowerCase()}</h2>
+          <StyledCopyButton onClick={handleCopyClipBoard} ref={copyButtonRef}>
+            복사
+          </StyledCopyButton>
+        </FlexDiv>
+        <div>
+          <DataGrid
+            getRowId={({ productId, type }) => `${productId}-${type}`}
+            rows={orderRows}
+            columns={productListColumns}
+            autoHeight
+            hideFooter
+            density="compact"
+            sx={{
+              borderLeft: "none",
+              borderRight: "none",
+            }}
+          />
+        </div>
+      </StyledModalContainer>
+    </StyledModal>
+  );
+};
+
+export default UserDetailModal;
