+++ src/components/Manager/User/UserTable.tsx
@@ -1,333 +1,11 @@
-import { Button } from "@mui/material";
-import { DataGrid, GridCellParams, GridColDef } from "@mui/x-data-grid";
+import { DataGrid, GridCellParams } from "@mui/x-data-grid";
 import { useOverlay } from "@toss/use-overlay";
-import { Dispatch, SetStateAction, useRef } from "react";
-import { styled } from "styled-components";
+import { Dispatch, SetStateAction } from "react";
 
-import { StyledModal } from "@/components/Manager/DashboardItems";
+import { tableColumns, UserTableRow } from "@/components/Manager/User/const";
+import UserDetailModal from "@/components/Manager/User/UserDetailModal";
+import { handleUserInfoListForTable } from "@/components/Manager/User/util";
 import { OrdererInfo } from "@/const";
-
-const tableColumns: GridColDef[] = [
-  { field: "id", headerName: "ID", width: 120 },
-  { field: "name", headerName: "Name", width: 107 },
-  {
-    field: "company",
-    headerName: "Company",
-    width: 107,
-  },
-  {
-    field: "email",
-    headerName: "E-mail",
-    width: 200,
-  },
-  {
-    field: "contactNumber",
-    headerName: "Contact Number",
-    width: 150,
-  },
-  {
-    field: "submissionTime",
-    headerName: "Submission Time",
-    width: 150,
-  },
-  {
-    field: "type",
-    headerName: "Type",
-    width: 150,
-  },
-];
-
-const personalInfoColumns1: GridColDef[] = [
-  { field: "id", headerName: "ID", width: 300 },
-  { field: "name", headerName: "Name", width: 150 },
-  { field: "company", headerName: "Company", width: 150 },
-];
-
-const personalInfoColumns2: GridColDef[] = [
-  { field: "email", headerName: "E-mail", width: 300 },
-  { field: "contactNumber", headerName: "Contact Number", width: 150 },
-  { field: "businessType", headerName: "Business Type", width: 150 },
-];
-
-const companyAddressColumns: GridColDef[] = [
-  { field: "postalCode", headerName: "Postal Code", width: 150 },
-  { field: "address", headerName: "Address", width: 250 },
-  { field: "detailAddress", headerName: "Detail Address", width: 250 },
-];
-
-const shippingAddressColumns: GridColDef[] = [
-  { field: "postalCode", headerName: "Postal Code", width: 150 },
-  { field: "address", headerName: "Address", width: 250 },
-  { field: "detailAddress", headerName: "Detail Address", width: 250 },
-];
-
-const productListColumns: GridColDef[] = [
-  { field: "productId", headerName: "Product ID", width: 150 },
-  { field: "type", headerName: "구분", width: 250 },
-  { field: "quantity", headerName: "개수", width: 200 },
-];
-
-type UserTableRow = {
-  id: string;
-  name: string;
-  company: string;
-  email: string;
-  contactNumber: string;
-  submissionTime: string;
-  type: string;
-  __user_info__: OrdererInfo;
-};
-
-const handleUserInfoListForTable = (
-  userInfoList: OrdererInfo[]
-): UserTableRow[] => {
-  return (
-    userInfoList?.map((userInfo) => {
-      const {
-        personalInfo: {
-          name,
-          companyName,
-          contactInfo: { email, phoneNumber },
-          businessType,
-        },
-        submissionTime,
-        userId,
-      } = userInfo;
-
-      return {
-        id: userId,
-        name,
-        company: companyName,
-        email: email,
-        contactNumber: `${phoneNumber.countryCode}${phoneNumber.number}`,
-        submissionTime,
-        type: businessType,
-        __user_info__: userInfo,
-      };
-    }) ?? []
-  );
-};
-
-const StyledModalContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 740px;
-  height: 650px;
-  background-color: var(--color-white);
-  overflow: scroll;
-  padding: 30px;
-
-  > div {
-    width: 100%;
-  }
-
-  h4 {
-    margin: 10px 0;
-    color: var(--color-h4-secondary);
-  }
-`;
-
-const FlexDiv = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-const StyledCopyButton = styled(Button)`
-  height: 30px;
-`;
-
-const handleUserInfoForOrder = (
-  hopeProducts: [
-    {
-      colorCardQuantity: number;
-      productId: string;
-      sampleYardages: [
-        {
-          colorId: string;
-          colorName: string;
-          yardageQuantity: number;
-        }
-      ];
-    }
-  ]
-) => {
-  return hopeProducts
-    .map((product) => {
-      const { colorCardQuantity, productId, sampleYardages } = product;
-      const res: any[] = [];
-      if (colorCardQuantity > 0) {
-        res.push({
-          productId,
-          type: "Color Card",
-          quantity: colorCardQuantity,
-        });
-      }
-
-      sampleYardages.forEach((sy) => {
-        const { colorId, colorName, yardageQuantity } = sy;
-        res.push({
-          productId,
-          type: `S/Y ${colorId} ${colorName}`,
-          quantity: yardageQuantity,
-        });
-      });
-
-      return res;
-    })
-    .flat();
-};
-
-const UserInfoDetailModal = ({
-  isModalOpen,
-  onModalClose,
-  modalUserInfoData,
-}: {
-  isModalOpen: boolean;
-  onModalClose: () => void;
-  modalUserInfoData: OrdererInfo;
-}) => {
-  const {
-    userId,
-    submissionTime,
-    productLengthUnit,
-    hopeProducts,
-    personalInfo: {
-      name,
-      companyName,
-      businessType,
-      contactInfo: { email, phoneNumber },
-      companyAddress,
-      shippingAddress,
-    },
-  } = modalUserInfoData;
-
-  const orderRows = handleUserInfoForOrder(hopeProducts);
-  const copyButtonRef = useRef<HTMLButtonElement>(null);
-
-  const handleCopyClipBoard = async () => {
-    const text = hopeProducts.map((product) => product.productId).join("\n");
-    try {
-      await navigator.clipboard.writeText(text);
-      if (copyButtonRef.current) {
-        copyButtonRef.current.textContent = "복사됨";
-      }
-    } catch (e) {
-      alert("복사에 실패하였습니다");
-    }
-  };
-
-  return (
-    <StyledModal open={isModalOpen} onClose={onModalClose}>
-      <StyledModalContainer>
-        <h2>User Info - {submissionTime ?? ""}</h2>
-        <h4>Personal Info</h4>
-        <div>
-          <DataGrid
-            rows={[
-              {
-                id: userId,
-                name,
-                company: companyName,
-              },
-            ]}
-            columns={personalInfoColumns1}
-            autoHeight
-            hideFooter
-            density="compact"
-            sx={{
-              borderLeft: "none",
-              borderRight: "none",
-            }}
-          />
-        </div>
-        <div>
-          <DataGrid
-            getRowId={() => userId}
-            rows={[
-              {
-                email,
-                contactNumber: `${phoneNumber.countryCode}${phoneNumber.number}`,
-                businessType,
-              },
-            ]}
-            columns={personalInfoColumns2}
-            autoHeight
-            hideFooter
-            density="compact"
-            sx={{
-              borderLeft: "none",
-              borderRight: "none",
-            }}
-          />
-        </div>
-        <h4>Company Address</h4>
-        <div>
-          <DataGrid
-            getRowId={() => userId}
-            rows={[
-              {
-                postalCode: companyAddress.postalCode,
-                address: companyAddress.address,
-                detailAddress: companyAddress.detailAddress,
-              },
-            ]}
-            columns={companyAddressColumns}
-            autoHeight
-            hideFooter
-            density="compact"
-            sx={{
-              borderLeft: "none",
-              borderRight: "none",
-            }}
-          />
-        </div>
-        <h4>Shipping Address</h4>
-        <div>
-          <DataGrid
-            getRowId={() => userId}
-            rows={[
-              {
-                postalCode: shippingAddress.postalCode,
-                address: shippingAddress.address,
-                detailAddress: shippingAddress.detailAddress,
-              },
-            ]}
-            columns={shippingAddressColumns}
-            autoHeight
-            hideFooter
-            density="compact"
-            sx={{
-              borderLeft: "none",
-              borderRight: "none",
-            }}
-          />
-        </div>
-
-        <FlexDiv>
-          <h2>Order - {productLengthUnit.toLowerCase()}</h2>
-          <StyledCopyButton onClick={handleCopyClipBoard} ref={copyButtonRef}>
-            복사
-          </StyledCopyButton>
-        </FlexDiv>
-        <div>
-          <DataGrid
-            getRowId={({ productId, type }) => `${productId}-${type}`}
-            rows={orderRows}
-            columns={productListColumns}
-            autoHeight
-            hideFooter
-            density="compact"
-            sx={{
-              borderLeft: "none",
-              borderRight: "none",
-            }}
-          />
-        </div>
-      </StyledModalContainer>
-    </StyledModal>
-  );
-};
 
 const UserTable = ({
   userInfoList,
@@ -340,36 +18,34 @@ const UserTable = ({
   const overlay = useOverlay();
 
   return (
-    <>
-      <div style={{ height: "calc(100% - 60px)", width: "100%" }}>
-        <DataGrid
-          rows={tableRows}
-          columns={tableColumns}
-          initialState={{
-            pagination: {
-              paginationModel: { page: 0, pageSize: 100 },
-            },
-          }}
-          pageSizeOptions={[10, 20, 50, 100]}
-          checkboxSelection
-          onRowSelectionModelChange={(selectedList: string[]) => {
-            setSelectedUserList(selectedList);
-          }}
-          onCellClick={(cell: GridCellParams<UserTableRow>, e) => {
-            if (cell.field !== "__check__") {
-              e.stopPropagation();
-              overlay.open(({ isOpen, close }) => (
-                <UserInfoDetailModal
-                  isModalOpen={isOpen}
-                  onModalClose={close}
-                  modalUserInfoData={cell.row.__user_info__}
-                />
-              ));
-            }
-          }}
-        />
-      </div>
-    </>
+    <div style={{ height: "calc(100% - 60px)", width: "100%" }}>
+      <DataGrid
+        rows={tableRows}
+        columns={tableColumns}
+        initialState={{
+          pagination: {
+            paginationModel: { page: 0, pageSize: 100 },
+          },
+        }}
+        pageSizeOptions={[10, 20, 50, 100]}
+        checkboxSelection
+        onRowSelectionModelChange={(selectedList: string[]) => {
+          setSelectedUserList(selectedList);
+        }}
+        onCellClick={(cell: GridCellParams<UserTableRow>, e) => {
+          if (cell.field !== "__check__") {
+            e.stopPropagation();
+            overlay.open(({ isOpen, close }) => (
+              <UserDetailModal
+                isModalOpen={isOpen}
+                onModalClose={close}
+                modalUserInfoData={cell.row.__user_info__}
+              />
+            ));
+          }
+        }}
+      />
+    </div>
   );
 };
